Update CMakeLists.txt, ComplexCoreLegacyUUIDMapping.hpp, MergeColonies.cpp, and 13 more files
